requirements.txt
absl-py==0.9.0
astor==0.8.1
-e git+https://github.com/openai/baselines.git@ea25b9e8b234e6ee1bca43083f8f3cf974143998#egg=baselines
click==7.1.1
cloudpickle==1.3.0
future==0.18.2
gast==0.3.3
google-pasta==0.2.0
grpcio==1.28.1
gym==0.15.7
gym-retro==0.7.1
h5py==2.10.0
joblib==0.14.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.2.1
numpy==1.18.2
opencv-python==4.2.0.34
protobuf==3.11.3
pyglet==1.5.0
scipy==1.4.1
six==1.14.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
tqdm==4.45.0
Werkzeug==1.0.1
wrapt==1.12.1
